V31: Apply Ascend NPU inference patch and rollback device logic
- Apply .half() FP16 conversion for NPU inference:
  - pretrained.py: policy.half() after loading
  - modeling_act.py: model.half(), backbone.half(), img.half()
- Rollback device detection to use "cuda" (torch_npu maps internally)
- Remove explicit NPU device checks from utils.py
- Keep device as "cuda" in model config for Ascend compatibility

+++ operating_platform/policy/pretrained.py
@@ -125,8 +125,7 @@ class PreTrainedPolicy(nn.Module, HubMixin, abc.ABC):
 
         policy.to(config.device)
         policy.eval()
-        # policy = policy.half()
-        policy=policy
+        policy = policy.half()
         return policy
 
     @classmethod
